pyproject.toml
[project]
name = "pinglog"
version = "0.1.0"
description = "A Telegram-powered habit and activity logger"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-telegram-bot[callback-data,job-queue,webhooks]>=22.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pinglog"]

[project.scripts]
pinglog = "pinglog.cli.commands:main"
